requirements.txt
discord.py
python-dotenv
pytube
mysql-connector-python